+++ src/angular/dialog/dialog/dialog.ts
@@ -105,4 +105,8 @@ export class SbbDialog {
   public openOutput = internalOutputFromObservable(
     fromEvent<Event>(this.#element.nativeElement, 'open'),
   );
+
+  public escapeStrategy(): void {
+    return this.#element.nativeElement.escapeStrategy();
+  }
 }