+++ src/showcase/stories/angular/autocomplete-grid/autocomplete-grid/autocomplete-grid.stories.ts
@@ -161,7 +161,9 @@ const meta: Meta = {
   component: SbbAutocompleteGrid,
   parameters: {
     backgroundColor: (context: StoryContext) =>
-      context.args['negative'] ? 'var(--sbb-color-black)' : 'var(--sbb-color-white)',
+      context.args['negative']
+        ? 'var(--sbb-background-color-1-negative)'
+        : 'var(--sbb-background-color-1)',
   },
   argTypes,
   args,