+++ src/showcase/stories/angular/button/secondary-button-link/secondary-button-link.stories.ts
@@ -55,7 +55,9 @@ const meta: Meta = {
   parameters: {
     actions: { handles: ['click'] },
     backgroundColor: (context: StoryContext) =>
-      context.args['negative'] ? 'var(--sbb-color-iron)' : 'var(--sbb-color-white)',
+      context.args['negative']
+        ? 'var(--sbb-background-color-1-negative)'
+        : 'var(--sbb-background-color-1)',
   },
   argTypes,
   args,