+++ src/showcase/stories/angular/dialog/dialog-title/dialog-title.stories.ts
@@ -30,7 +30,9 @@ const meta: Meta = {
   component: SbbDialogTitle,
   parameters: {
     backgroundColor: (context: StoryContext) =>
-      context.args['negative'] ? 'var(--sbb-color-charcoal)' : 'var(--sbb-color-white)',
+      context.args['negative']
+        ? 'var(--sbb-background-color-1-negative)'
+        : 'var(--sbb-background-color-1)',
   },
   argTypes,
   args,