+++ src/showcase/stories/angular/link/link-button/link-button.stories.ts
@@ -61,7 +61,9 @@ const meta: Meta = {
   parameters: {
     actions: { handles: ['click'] },
     backgroundColor: (context: StoryContext) =>
-      context.args['negative'] ? 'var(--sbb-color-charcoal)' : 'var(--sbb-color-white)',
+      context.args['negative']
+        ? 'var(--sbb-background-color-1-negative)'
+        : 'var(--sbb-background-color-1)',
   },
   argTypes,
   args,
@@ -69,7 +71,7 @@ const meta: Meta = {
     props: { text, ...args },
     template: `
        <p
-        style="${args['negative'] ? 'color: var(--sbb-color-aluminium);' : 'color: var(--sbb-color-iron)'}"
+        style="${args['negative'] ? 'color: var(--sbb-color-4-negative);' : 'color: var(--sbb-color-4)'}"
         class="sbb-text-m"
       >
         Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt