+++ src/showcase/stories/angular/teaser-product/teaser-product/teaser-product.stories.ts
@@ -73,7 +73,9 @@ const meta: Meta = {
   component: SbbTeaserProduct,
   parameters: {
     backgroundColor: (context: StoryContext) =>
-      context.args['negative'] ? 'var(--sbb-color-black)' : 'var(--sbb-color-white)',
+      context.args['negative']
+        ? 'var(--sbb-background-color-1-negative)'
+        : 'var(--sbb-background-color-1)',
   },
   argTypes,
   args,
@@ -82,7 +84,7 @@ const meta: Meta = {
     template: `
       <sbb-teaser-product ${argsToTemplate({ imageAlignment, ...args })}>
         <figure slot="image" class="sbb-figure">
-          <img src='https://cdn.img.sbb.ch/content/dam/internet/lyne/Hoehenrundweg-Gryden-Lenk.jpg' alt="" />
+          <img src='https://cdn.img.sbb.ch/content/dam/internet/externe-assets/lyne/Hoehenrundweg-Gryden-Lenk.jpg' alt="" />
           ${
             withChip
               ? `<sbb-chip-label class=${imageAlignment === 'after' ? 'sbb-figure-overlap-start-end' : 'sbb-figure-overlap-start-start'}>